requirements.txt
langchain
langchain-openai
langchain-community
pypdf
python-dotenv
pillow
lxml
tiktoken
langchain-pinecone
redis
"unstructured[all-docs]"